refactor(tests): relocate the completion test into cli.rs (#224)
Move `test_completion_generation_succeeds_for_all_shells` out of lib.rs's
`mod tests` into a new `#[cfg(test)] mod tests` in cli.rs, next to the `Cli`
it exercises (extracted in #233). This is the last of the #217-extraction
tests left behind in lib.rs — paths / merge / url / plugin_build landed in
#248.

Pure relocation, no assertion changes; the test keeps its local
`use clap::CommandFactory` / `use clap_complete::Shell` and resolves `Cli`
through `use super::*`. cargo make check green (822 tests, plural doctest,
clippy -D warnings --all-targets, fmt, lock-check).

Closes #224. Refs #217.
